+++ css/dados.css
@@ -1,5 +1,5 @@
 body {
-    background-image: url('../images/ilha-luanda.jpg');
+    background-image: url('../assets/images/ilha-luanda.jpg');
     background-size: cover;
     background-repeat: no-repeat;
     font-family: 'Gill Sans', 'Gill Sans MT', Calibri, 'Trebuchet MS', sans-serif;
