+++ index.html
@@ -6,34 +6,34 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>NÃO TE IRRITES</title>
-    <link rel="stylesheet" href="./assets/css/dados.css">
-    <link rel="stylesheet" href="./assets/css/style.css">
+    <link rel="stylesheet" href="./css/dados.css">
+    <link rel="stylesheet" href="./css/style.css">
 </head>
 
 <body>
     <div class="nanter-container">
         <div class="nanter">
             <div class="pecas">
-                <div class="peca" PID="P1" peca="0"></div>
-                <div class="peca" PID="P1" peca="1"></div>
-                <div class="peca" PID="P1" peca="2"></div>
-                <div class="peca" PID="P1" peca="3"></div>
+                <div class="peca" PID="P1" peca="0" id="1"></div>
+                <div class="peca" PID="P1" peca="1" id="1"></div>
+                <div class="peca" PID="P1" peca="2" id="1"></div>
+                <div class="peca" PID="P1" peca="3" id="1"></div>
 
-                <div class="peca" PID="P2" peca="0"></div>
-                <div class="peca" PID="P2" peca="1"></div>
-                <div class="peca" PID="P2" peca="2"></div>
-                <div class="peca" PID="P2" peca="3"></div>
+                <div class="peca" PID="P2" peca="0" id="2"></div>
+                <div class="peca" PID="P2" peca="1" id="2"></div>
+                <div class="peca" PID="P2" peca="2" id="2"></div>
+                <div class="peca" PID="P2" peca="3" id="2"></div>
 
 
-                <div class="peca" PID="P3" peca="0"></div>
-                <div class="peca" PID="P3" peca="1"></div>
-                <div class="peca" PID="P3" peca="2"></div>
-                <div class="peca" PID="P3" peca="3"></div>
+                <div class="peca" PID="P3" peca="0" id="3"></div>
+                <div class="peca" PID="P3" peca="1" id="3"></div>
+                <div class="peca" PID="P3" peca="2" id="3"></div>
+                <div class="peca" PID="P3" peca="3" id="3"></div>
 
-                <div class="peca" PID="P4" peca="0"></div>
-                <div class="peca" PID="P4" peca="1"></div>
-                <div class="peca" PID="P4" peca="2"></div>
-                <div class="peca" PID="P4" peca="3"></div>
+                <div class="peca" PID="P4" peca="0" id="4"></div>
+                <div class="peca" PID="P4" peca="1" id="4"></div>
+                <div class="peca" PID="P4" peca="2" id="4"></div>
+                <div class="peca" PID="P4" peca="3" id="4"></div>
             </div>
 
             <div class="player-bases">
@@ -126,12 +126,12 @@
                     </div>
                 </div>
 
-                <button id="reset-btn" class="btn btn-reset">Repetir</button>
+                <button id="reset-btn" class="btn btn-reset">Novo Jogo</button>
             </div>
             <h2 class="active-player" hidden> <span></span> </h2>
         </div>
     </div>
-    <script src="./index.js" type="module"></script>
+    <script src="./js/index.js" type="module"></script>
     <script src="./sweetalert2.all.min.js"></script>
 </body>
