--- index.js
@@ -1,3 +1,0 @@
-
-import { Ludo } from './model/Ludo.js';
-const ludo = new Ludo();
