+++ js/index.js
@@ -1,0 +1,3 @@
+
+import { Nanter } from './Nanter.js';
+const nanter = new Nanter();
